コインの種類を大幅に追加。12桁くらいまで耐えるようにした。

+++ CEffect_MoneyNumber.h
@@ -2,13 +2,13 @@
 #include "CEffect.h"
 class CEffect_MoneyNumber : public CEffect {
  public:
-  CEffect_MoneyNumber(CVector position_, int num);
+  CEffect_MoneyNumber(CVector position_, double num);
 
   void Update();
   void Render() const;
 
  private:
-  int num_;
+  double num_;
   double dy_;
   int cnt_;
 };
